[refactor] renamed crypto to hash and added it to primitives
