+++ bscript/address.go
@@ -6,8 +6,8 @@ import (
 	"fmt"
 
 	base58 "github.com/bitcoin-sv/go-sdk/compat/base58"
-	"github.com/bitcoin-sv/go-sdk/crypto"
 	ec "github.com/bitcoin-sv/go-sdk/primitives/ec"
+	crypto "github.com/bitcoin-sv/go-sdk/primitives/hash"
 	"github.com/bitcoin-sv/go-sdk/util"
 )
