+++ compat/bip32/extendedkey.go
@@ -21,8 +21,8 @@ import (
 
 	"github.com/bitcoin-sv/go-sdk/chaincfg"
 	base58 "github.com/bitcoin-sv/go-sdk/compat/base58"
-	"github.com/bitcoin-sv/go-sdk/crypto"
 	ec "github.com/bitcoin-sv/go-sdk/primitives/ec"
+	crypto "github.com/bitcoin-sv/go-sdk/primitives/hash"
 )
 
 const (
