+++ compat/wif/wif.go
@@ -10,8 +10,8 @@ import (
 
 	"github.com/bitcoin-sv/go-sdk/chaincfg"
 	base58 "github.com/bitcoin-sv/go-sdk/compat/base58"
-	"github.com/bitcoin-sv/go-sdk/crypto"
 	ec "github.com/bitcoin-sv/go-sdk/primitives/ec"
+	crypto "github.com/bitcoin-sv/go-sdk/primitives/hash"
 )
 
 var (
