+++ primitives/hash/hash.go
@@ -1,4 +1,4 @@
-package crypto
+package primitives
 
 import (
 	"crypto/hmac"
