+++ primitives/hash/hash_test.go
