+++ util/bytemanipulation_test.go
@@ -4,7 +4,7 @@ import (
 	"encoding/hex"
 	"testing"
 
-	"github.com/bitcoin-sv/go-sdk/crypto"
+	crypto "github.com/bitcoin-sv/go-sdk/primitives/hash"
 
 	"github.com/bitcoin-sv/go-sdk/util"
 	"github.com/stretchr/testify/assert"
